% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barplot_by_groups}
\alias{barplot_by_groups}
\title{Draw a bar plot of counts by group}
\usage{
barplot_by_groups(df, value = "rs_id", group.by)
}
\arguments{
\item{df}{A data frame with at least two columns.}

\item{value}{Value to count.}

\item{group.by}{Column by which counts will be grouped.}
}
\value{
A bar plot representing value counts across groups.
}
\description{
\code{barplot_by_groups} takes a data frame, a value to count, and a category
by which to group the counts. Default is counting number of variants across
groups.
}
\examples{
barplot_example_df <- data.frame(rs_id=rep('rs7357', 10),
								relevance=c(rep('important', 2),
										rep('not_important', 8)))
barplot_by_groups(barplot_example_df, "rs_id", "relevance")
}
