% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAGGER.R
\name{DAGGER}
\alias{DAGGER}
\title{Wrapper for default workflow}
\usage{
DAGGER(GWAS, GTEx, DGIdb)
}
\arguments{
\item{GWAS, GTEx, DGIdb}{GWAS, gene expression and drug-gene interaction
dataframes.}
}
\value{
A data frame containing full repositioning analysis. Keeps all
colums from input data.
}
\description{
\code{DAGGER} calls \code{merge_gene_var_drug}, \code{predict_effect} and \code{get_candidates},
thus running the default workflow. This simple structure is the most
bare-bones case, but is effective nonetheless. Resulting object may be
further processed for more specific results.
}
\examples{
GWAS_example <- GWAS_demo[501:511, ]
GTEx_example <- GTEx[c(1:5, 1598, 146580, 219755), ]
DGIdb_example <- DGIdb[c((10:14), 5307, 5853, 8511), ]
DAGGER_result <- DAGGER(GWAS_example, GTEx_example, DGIdb_example)
}
