% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_filter.R
\name{filter_significance}
\alias{filter_significance}
\title{Filter dataframe by p-value}
\usage{
filter_significance(df, value = 0.05)
}
\arguments{
\item{df}{A data frame.}

\item{value}{P-value cutoff.}
}
\value{
A subset of the original data frame with all rows passing filter.
}
\description{
\code{filter_significance} filters out all rows in a data frame below input
p-value (default 0.05).
}
\examples{
example_df <- head(GWAS_demo)
colnames(example_df)[2] <- "p_value"
example_df <- filter_significance(example_df, 1e-20)
}
