% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositioning.R
\name{get_candidates}
\alias{get_candidates}
\title{Filter potential candidates for drug repositioning}
\usage{
get_candidates(df)
}
\arguments{
\item{df}{A data frame with at least two columns: interaction_types and
prediction.}
}
\value{
Input dataset with an additional column recommending candidates
for drug repositioning.
}
\description{
\code{get_candidates} compares predicted beneficial drug effect with described
gene-drug interaction. If they match, row is marked as potential candidate.
}
\examples{
repositioning_example <- data.frame(interaction_types = c("Missing",
								"inhibitor", "agonist", "activator",
							"positive modulator", "partial agonist",
								"inducer", "allosteric modulator",
							"Missing", "activator", "inhibitor",
							"blocker", "antagonist",
							"inverse agonist", "negative modulator",
							"antisense oligonucleotide", "suppressor",
							"inhibitory allosteric modulator"),
							prediction = c(rep("activator", 8),
								rep("inhibitor", 10)))
get_candidates(repositioning_example)
}
