% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GWAS_demo}
\alias{GWAS_demo}
\title{GWAS demo dataset}
\format{
\subsection{\code{GWAS_demo}}{

A data frame with 6 rows and 3 columns:
\describe{
\item{rs_id}{Reference SNP ID}
\item{pvalue}{Statistical significance of SNP association to AD risk}
\item{beta_number}{Beta coefficient of SNP association to AD. Positive}
}
}
}
\source{
\url{https://www.ebi.ac.uk/gwas/studies/GCST90245848}
}
\usage{
GWAS_demo
}
\description{
A height Genome-Wide Association Study taken from The GWAS
Catalog using the gwasrapidd package
}
\keyword{datasets}
