% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_variants.R
\name{locate_variants}
\alias{locate_variants}
\title{Locate variants in genome, classifying them in intragenic and intergenic}
\usage{
locate_variants(rs_list)
}
\arguments{
\item{df}{A data frame with at least two columns.}

\item{col1, col2}{Columns containing categories to test.}

\item{col1_val1, col1_val2, col2_val1, col2_val2}{Values of the first and
second columns, respectively, to test.}
}
\value{
A list containing the chi2 table and test results, with chi2 value,
p-value and effect size (Cramer-V)
}
\description{
\code{test_chi2} takes a data frame, two column names, and two values for each
column and formats the data into a chi2-test compatible format. Then,
it performs the chi2 statistical tests, including effect size (Cramer-V)
calculation.
}
\examples{
chisq_example <- data.frame(Tissue=c("Brain", "Brain", "Stomach"),
								candidate=c(TRUE, TRUE, FALSE))
chisq_example <- chisq_example[rep(seq_len(nrow(chisq_example)),
								each = 20), ]
test_chi2(chisq_example, "Tissue", "brain", "stomach", "candidate", T, F)
}
