% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gene_var_drug.R
\name{merge_gene_var_drug}
\alias{merge_gene_var_drug}
\title{Merge expression, GWAS and gene-drug interaction databases}
\usage{
merge_gene_var_drug(GWAS, GTEx, DGIdb)
}
\arguments{
\item{GWAS, GTEx, DGIdb}{GWAS, gene expression and drug-gene interaction
dataframes.}
}
\value{
Merged dataframe.
}
\description{
\code{merge_gene_var_drug} takes three dataframes, one of GWAS, one of gene
expression and one of gene-drug interaction information and merges them.
It calls \code{remove_duplicate_rs}, \code{filter_significance} and
\code{parse_column_names} as needed. GWAS and GTEx dataframes are merged
by rs_id column, result is merged with DGIdb by gene_symbol column'
}
\examples{
GWAS_example <- GWAS_demo[501:511, ]
GTEx_example <- GTEx[c(1:5, 1598, 146580, 219755), ]
DGIdb_example <- DGIdb[c((10:14), 5307, 5853, 8511), ]
merged <- merge_gene_var_drug(GWAS_example, GTEx_example, DGIdb_example)
}
