% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot of variant p-value versus odds ratio}
\usage{
plot_volcano(
  df,
  title = "Odds Ratio vs Variant p-value",
  or_cutoff = 1.05,
  pval_cutoff = 0.001,
  fontsize = 32
)
}
\arguments{
\item{df}{A DAGGER-parsed data frame containing p-value and beta number
columns.}

\item{title}{Title of the resulting plot.}

\item{or_cutoff}{Value from which statistical significance line position for
odds ratio (X axis) will be calculated. Symmetrical line will be drawn in
opposite value in X-axis (log2 of odds-ratio distribution is symmetrical).}

\item{pval_cutoff}{Value from which statistical significance line position
for p-value (Y axis) will be calculated.}

\item{fontsize}{Base font size for text. Titles will be bigger, axes
will be lower.}
}
\value{
A vulcano plot of the provided distributions.
}
\description{
\code{plot_volcano} plots log2 of variant odds ratio versus log10 of p-value
for a given DAGGER-parsed dataframe containing variant data.
}
\examples{
GWAS_example <- GWAS_demo
colnames(GWAS_example) <- c("rs_id", "p_value", "beta_number")
plot_volcano(GWAS_example)
}
