% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositioning.R
\name{predict_effect}
\alias{predict_effect}
\title{Predict beneficial drug effect with variant risk and expression effect}
\usage{
predict_effect(gene_variant_df)
}
\arguments{
\item{gene_variant_df}{a dataframe resulted from merging a DAGGER-parsed
variant-risk dataframe and a DAGGER-parsed gene-variant dataframe.}
}
\value{
Input dataframe with an additional column with the predicted
beneficial drug effect according to beta and slope columns.
}
\description{
\code{predict_effect} compares beta and slope columns in DAGGER dataframe
(containing variant and gene expression information) and predicts
beneficial drug effect.
}
\section{Logic behind the function}{

\itemize{
\item Negative betas mean protection
\item Positive betas mean risk.
\item Negative slopes mean lower expression
\item Positive slopes mean higher expression.
\item If signs are opposite (prediction == TRUE), either the risk
variant decreases expression or the protective variant increases it,
therefore an activator could be beneficial.
\item If signs are equal (prediction == FALSE), either the protective variant
decreases expression or the risk variant increases it. Either way,
an inhibitor would be appropriate.
}
}

\examples{
example_df <- merge_gene_var_drug(GWAS_demo, GTEx, DGIdb)[10:20, ]
example_df <- predict_effect(example_df)
}
