% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_filter.R
\name{remove_duplicate_rs}
\alias{remove_duplicate_rs}
\title{Remove duplicate RS in GWAS dataframe}
\usage{
remove_duplicate_rs(df)
}
\arguments{
\item{df}{A DAGGER-parsed data frame of GWAS information. Must contain an
rs_id column. p_value column optional, but recommended.}
}
\value{
A data frame with no duplicate RS. Only the first occurrence of each
duplicate will remain, or the most statistically significant if p-value
column is found'
}
\description{
\code{remove_duplicate_rs} removes rows corresponding to duplicate rs, keeping
the first occurrence. Dataframe is first sorted by p-value if column exists.
}
\examples{
dupes_example <- data.frame(
					rs_id = c("rs2710888", "rs182532", "rs9660106"),
						p_val_variant = c(2e-58, 1e-17, 2e-12),
						beta_number = GWAS_demo$beta_number[c(1, 3, 5)]
													+ 0.002)
dupes_example_b <- head(GWAS_demo)
colnames(dupes_example_b)[1:2] <- c("rs_id", "p_value")
dupes_example <- rbind(dupes_example, dupes_example_b)
dupes_example <- remove_duplicate_rs(dupes_example)
}
