% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GTEx}
\alias{GTEx}
\title{GTEx dataset}
\format{
\subsection{\code{GTEx}}{

A data frame with 1208024 rows and 9 columns:
\describe{
\item{Tissue}{Tissue where gene expression is altered}
\item{gene_id}{Ensembl ID of affected gene}
\item{gene_name}{GENCODE gene name}
\item{variant_information}{Unique variant identifier. Contains chromosome,
position, ref and alt alleles and genome build. Named "variant_id" in
original GTEx dataset'}
\item{tss_distance}{Distance to gene transcription start site}
\item{rs_id_dbSNP151_GRCh38p7}{Reference SNP ID}
\item{slope}{Effect on gene expression}
\item{pval_nominal}{statistical significance of gene-variant association}
}
}
}
\source{
\url{https://www.gtexportal.org/home/downloads/adult-gtex#qtl}
}
\usage{
GTEx
}
\description{
Merged contents GTEx_Analysis_v8_eQTL.tar, shows variant-gene expression
association. Processing detailed under data-raw/, files GTEx.sh and GTEx.R
}
\keyword{datasets}
