% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GWAS_demo}
\alias{GWAS_demo}
\title{GWAS demo dataset}
\format{
\subsection{\code{GWAS_demo}}{

A data frame with 6 rows and 3 columns:
\describe {
\item{variant_id}{Reference SNP ID}
\item{pvalue}{Statistical significance of SNP association to AD risk}
\item{beta_number}{Beta coefficient of SNP association to AD. Positive}
}
}
}
\source{
<Detailed in data-raw/GWAS_demo.R>
}
\usage{
GWAS_demo
}
\description{
A very small Alzheimer's Diseasevariant risk association taken from The GWAS
Catalog via the gwasrapidd package
}
\keyword{datasets}
